requirements.txt
torch
torchvision
timm
gradio==4.29.0
Pillow
scikit-learn
numpy
matplotlib
